BLAST Top Hit

Conduct BLAST search and return top hit

blast_top_hit(
  query,
  db,
  type = "blastn",
  identity = 95,
  coverage = 95,
  evalue = 1e+06,
  max_target_seqs = 5,
  max_hsp = 5,
  ranks = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  delim = ";",
  tie = "first",
  args = NULL,
  quiet = FALSE,
  ...
)

Arguments

query

(Required) Query sequence. Accepts a DNABin object, DNAStringSet object, Character string, or filepath.

db

(Required) Reference sequences to conduct search against. Accepts a DNABin object, DNAStringSet object, Character string, or filepath. If DNAbin, DNAStringSet or character string is provided, a temporary fasta file is used to construct BLAST database

type

(Required) type of search to conduct, default 'blastn'

identity

(Required) Minimum percent identity cutoff.

coverage

(Required) Minimum percent query coverage cutoff.

evalue

(Required) Minimum expect value (E) for saving hits

max_target_seqs

(Required) Number of aligned sequences to keep. Even if you are only looking for 1 top hit keep this higher for calculations to perform properly.

max_hsp

(Required) Maximum number of HSPs (alignments) to keep for any single query-subject pair.

ranks

(Required) The taxonomic ranks contained in the fasta headers

delim

(Required) The delimiter between taxonomic ranks in fasta headers

tie

How to handle ties in top hit results. Options are to break ties by selecting the first hit (Default), or return all tied hits.

args

(Optional) Extra arguments passed to BLAST

quiet

(Optional) Whether progress should be printed to console, default is FALSE

Value