Fetch sequences from BOLD

Fetch sequences from BOLD

fetch_bold(
  x,
  marker = "COI-5P",
  output = "gb-binom",
  quiet = FALSE,
  db = NULL,
  retry_attempt = 3,
  retry_wait = 5
)

Arguments

x

A taxon name or vector of taxa to download sequences for

marker

the barcode marker used as a search term for the database

output

the output format for the taxonomy in fasta headers. Options include "h" for full heirarchial taxonomy (Accession;Domain;Phylum;Class;Order;Family;Genus;Species), "binom" for just genus species binomials (Accession;Genus_species), "bold" for BOLD taxonomic ID only (Accession;BoldTaxID), "gb" for genbank taxonomic ID (Accession|GBTaxID), "gb-binom" which outputs genbank taxonomic ID's and Genus species binomials, translating BOLD taxonomic ID's to genbank in the process (Accession|GBTaxID;Genus_species) or "standard" which outputs the default format for each database. For genbank this is the description field, and for bold this is sampleid|species name|markercode|genbankid

quiet

Whether progress should be printed to the console.

db

(Optional) a database file generated using taxreturn::get_ncbi_taxonomy() or taxreturn::get_ott_taxonomy()

retry_attempt

The number of query attempts in case of query failure due to poor internet connection.

retry_wait

How long to wait between query attempts

Value