Get subalignment

Aligns a DNABin to a reference PHMM model, and returns the optimal path

get_subalignment(
  x,
  model,
  tryrc = FALSE,
  quiet = FALSE,
  check_indels = TRUE,
  min_score = 10,
  omit_endgaps = FALSE,
  multithread = FALSE,
  ...
)

Arguments

x
A DNAbin object or an object coercible to DNAbin

model
A profile hidden Markov model (a "PHMM" object) generated by the aphid R package to align the sequences to.

tryrc
Whether the reverse complement should also be aligned

quiet
Whether progress should be printed to the console.

check_indels
Check that indels are multiples of 3, recommended for coding sequences such as COI

min_score
Minimum score for the viterbi alignment

omit_endgaps
Should gap characters at each end of the sequences be ignored when deriving the transition probabilities of the model? Defaults to FALSE. Set to TRUE if x is not a strict global alignment (i.e. if the alignment contains partial sequences with missing sections represented with gap characters).

multithread
Whether multithreading should be used, if TRUE the number of cores will be automatically detected, or provided a numeric vector to manually set the number of cores to use

...
aditional arguments to be passed to "Viterbi"

Value