Train IDTAXA

Train IDTAXA

train_idtaxa(
  x,
  max_group_size = 10,
  max_iterations = 3,
  allow_group_removal = TRUE,
  orient = FALSE,
  get_lineage = FALSE,
  db = NULL,
  quiet = FALSE
)

Arguments

x
A DNAbin

max_group_size
The maximum size of any taxonomic group. This be set to Inf (infinity) to allow for an unlimited number of sequences per group.

max_iterations
The number of iterations to conduct training for in order to identify any training sequences whose assigned classifications completely disagree with their predicted classification

allow_group_removal
Whether sequences that are the last remaining representatives of an entire group in the training data can be removed This can occur if the entire group appears to be misplaced in the taxonomic tree.

orient
Training sequences must all be in the same orientation. Set this to TRUE to reorient the sequences if you are unsure.

get_lineage
Get full taxonomic lineage using reformat_hierarchy if not already present

db
A database file generated by get_ncbi_taxonomy or get_ott_lineage. Required if get_lineage is TRUE

quiet
Whether progress should be printed to console

Value